backend/requirements.txt
# ArXiv RAG Assistant - Mechanistic Interpretability
# Install: pip install -r requirements.txt

# ML / Models
torch>=2.0.0
sentence-transformers>=2.2.2
transformers>=4.34.0
accelerate>=0.20.3
qdrant-client>=1.9.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
google-genai>=2.1.0
groq>=0.9.0

# ArXiv ingestion & PDF extraction
requests>=2.31.0
lxml>=4.9.0
feedparser>=6.0.0
PyMuPDF>=1.24.0
pdfplumber>=0.10.0

# Database (PostgreSQL)
psycopg[binary]>=3.1.0

# Tokenization
tiktoken>=0.5.0

# Image processing
Pillow>=10.0.0

# Citation expansion
semanticscholar>=0.8.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
numpy>=1.24.0
tqdm>=4.65.0
boto3>=1.34.0
nltk>=3.8.0
joblib>=1.3.0
rank_bm25>=0.2.2
psutil>=5.9.0
click>=8.1.7

# Evaluation (RAGAS)
ragas>=0.2.0
langchain-google-genai>=1.0.4
langchain-openai>=0.2.0

# Testing
pytest>=7.4.0
httpx>=0.25.0

# Auth & Security
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Async Database (Application DB — Supabase)
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0

# Redis Caching
redis>=5.0.0

# Email validation
email-validator>=2.0.0
